Update PieRadarChartBase.java
The first comment is that when we draw the pie chart inside mCircleBox, it is the center of this object that should be considered for finding distance to points or angles. The second major is that in finding an angle, the arc tan function is a standard and straightforward way.